requirements.txt
requests>=2.31.0
boto3>=1.28.0
PyQt6>=6.5.0
cryptography>=41.0.0
python-dotenv>=1.0.0
